Updated 'fpl_stats.xlsm'
Code improvements to make 'mainModule' macro more concise and streamline. Moved 'convertColumn' function to the end of 'mainModule' pipeline. This will mean the 'convertColumn' function will be applied to less cells. The reason this is good is because I've noticed a massive slowdown in performance when using this function. The rest of the code, overall, remains the same for other modules.